Commit backup player data, remove shared allowlist
Added a "canonical" copy of the main database to versioning, which is however not updated automatically.
(Has to be done manually, should be done periodically.)

This is implemented so every contributor can populate the main allowlist/database at data/playerNames.json with their own accounts set as admin/bot users for development/testing.